Combine mixins and sqn to utils and add test
